third_party/blink/web_tests/platform/generic/fast/peerconnection/RTCPeerConnection-stats-expected.txt

Tests the RTCPeerConnection stats interface.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pc = new RTCPeerConnection() did not throw exception.
PASS pc.getStats(statsHandler1) did not throw exception.
PASS statsHandler1 was called
PASS status is non-null.
PASS result.length is >= 0
PASS getUserMedia({audio:true, video:true}, gotStream) did not throw exception.
PASS Got a stream.
PASS pc.getStats(statsHandler2) did not throw exception.
PASS statsHandler2 was called
PASS result.length is >= 2
PASS timediff is >= 0
PASS res.id is non-null.
PASS res.type is non-null.
PASS res.names().length is >= 1
PASS status_g.namedItem(res.id) is non-null.
PASS status_g[res.id] is non-null.
PASS pc.close() did not throw exception.
PASS pc.getStats(statsHandler3) did not throw exception.
PASS statsHandler3 was called
PASS result.length is >= 2
PASS timediff is >= 0
PASS res.id is non-null.
PASS res.type is non-null.
PASS res.names().length is >= 1
PASS status_g.namedItem(res.id) is non-null.
PASS status_g[res.id] is non-null.
PASS pc.getStats(statsHandler4, 42) did not throw exception.
PASS statsHandler4 was called
PASS result.length is >= 2
PASS timediff is >= 0
PASS res.id is non-null.
PASS res.type is non-null.
PASS res.names().length is >= 1
PASS status_g.namedItem(res.id) is non-null.
PASS status_g[res.id] is non-null.
PASS successfullyParsed is true

TEST COMPLETE
